make species a generic string on the server
Because the client was written to only expect Cat or Dog, the server and client become out of sync.

The client can send a request to the server with an unknown species:
> Try to send an animal with an unknown species...
> 200

But when the server contains an animal with an unknown species, the client fails to parse the animals:
> Try to get all animals with an unknown species...
> 200
> Error at animal 0:
> ---> Error parsing `.species`:
> ---> ---> Expected one of: Dog | Cat
> ---> ---> But got: Robot

As a result, this is a breaking change for the client. The client can either:
1) Loosen the types locally
2) Ignore unknown animals

Let's look at both.